+++ include/Class_Para_Tree_3D.tpp
@@ -4560,7 +4560,7 @@ public:
 			}
 			updateAdapt();
 			setPboundGhosts();
-			log.writeLog(" Number of octants after Refine	:	" + to_string(global_num_octants));
+			log.writeLog(" Number of octants after Refine	:	" + to_string(static_cast<unsigned long long>(global_num_octants)));
 			nocts = octree.getNumOctants();
 
 			// Coarse
@@ -4678,7 +4678,7 @@ public:
 				localDone = true;
 			updateAdapt();
 			setPboundGhosts();
-			log.writeLog(" Number of octants after Refine	:	" + to_string(global_num_octants));
+			log.writeLog(" Number of octants after Refine	:	" + to_string(static_cast<unsigned long long>(global_num_octants)));
 			nocts = octree.getNumOctants();
 
 			// Coarse
